+++ CHANGELOG.md
@@ -23,6 +23,7 @@ at anytime.
   * `file_list` for files with bad signatures
   * return None from resolve commands when nothing is found
   * return lbry files with claims that are abandoned
+  * unhelpful error messages in `publish` and `channel_new`
 
 ## [0.9.2rc9] - 2017-04-08
 ### Added